Fixed product quantity not resetting on page change
